align audit logs response total count type with paging model

+++ src/modules/Audit/Audit.Application/Contracts/Responses/GetAuditLogsResponse.cs
@@ -8,7 +8,7 @@ public sealed class GetAuditLogsResponse
 
     public int PageSize { get; init; }
 
-    public long TotalCount { get; init; }
+    public int TotalCount { get; init; }
 
     public int TotalPages { get; init; }
 }
